refactor(ms-bank-account): changing the location of some directories
