+++ bank-account/src/main/java/com/dac/bank_account/command/repository/AccountCommandRepository.java
@@ -3,7 +3,9 @@ package com.dac.bank_account.command.repository;
 import com.dac.bank_account.command.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface AccountCommandRepository extends JpaRepository<Account, Long> {
-    Account findByAccountNumber(String accountNumber);
+    Optional<Account> findByAccountNumber(String accountNumber);
     boolean existsByAccountNumber(String accountNumber);
 }
