+++ bank-account/src/main/java/com/dac/bank_account/command/service/AccountCommandService.java
@@ -6,8 +6,7 @@ import com.dac.bank_account.command.dto.response.AccountResponseDTO;
 import com.dac.bank_account.command.dto.response.MovementResponseDTO;
 import com.dac.bank_account.command.dto.response.TransferResponseDTO;
 import com.dac.bank_account.command.entity.Account;
-import com.dac.bank_account.command.enums.TransactionType;
-import com.dac.bank_account.command.events.*;
+import com.dac.bank_account.enums.TransactionType;
 import com.dac.bank_account.command.repository.AccountCommandRepository;
 import org.springframework.stereotype.Service;
 
@@ -40,59 +39,54 @@ public class AccountCommandService {
     }
 
     public MovementResponseDTO deposit(String accountNumber, BigDecimal amount) {
-        Account account = accountCommandRepository.findByAccountNumber(accountNumber);
-        if (account == null) {
-            throw new IllegalArgumentException("Account not found with account number: " + accountNumber);
-        }
+        Account account = accountCommandRepository.findByAccountNumber(accountNumber)
+                        .orElseThrow(() -> new IllegalArgumentException("Account not found with account number: " + accountNumber));
         account.deposit(amount);
 
-        var mov = accountMapper.toEntity(account.getAccountNumber(), TransactionType.DEPOSITO, amount, null);
-        account.getTransactions().add(mov);
+        var transaction = accountMapper.toEntity(account.getAccountNumber(), TransactionType.DEPOSITO, amount, null);
+        account.getTransactions().add(transaction);
         accountCommandRepository.save(account);
 
         return new MovementResponseDTO(
                 account.getAccountNumber(),
-                LocalDateTime.now(ZoneOffset.of("-03:00")).toString(),
+                OffsetDateTime.now(ZoneOffset.of("-03:00")).toString(),
                 account.getBalance()
         );
     }
 
     public MovementResponseDTO withdraw(String accountNumber, BigDecimal amount) {
-        Account account = accountCommandRepository.findByAccountNumber(accountNumber);
-        if (account == null) {
-            throw new IllegalArgumentException("Account not found with account number: " + accountNumber);
-        }
+        Account account = accountCommandRepository.findByAccountNumber(accountNumber)
+                        .orElseThrow(() -> new IllegalArgumentException("Account not found with account number: " + accountNumber));
         account.withdraw(amount);
 
-        var mov = accountMapper.toEntity(account.getAccountNumber(), TransactionType.SAQUE, amount, null);
-        account.getTransactions().add(mov);
+        var transaction = accountMapper.toEntity(account.getAccountNumber(), TransactionType.SAQUE, amount, null);
+        account.getTransactions().add(transaction);
         accountCommandRepository.save(account);
 
         return new MovementResponseDTO(
                 account.getAccountNumber(),
-                LocalDateTime.now(ZoneOffset.of("-03:00")).toString(),
+                OffsetDateTime.now(ZoneOffset.of("-03:00")).toString(),
                 account.getBalance()
         );
     }
 
     public TransferResponseDTO transfer(String sourceAccountNumber, BigDecimal amount, String targetAccountNumber) {
-        Account source = accountCommandRepository.findByAccountNumber(sourceAccountNumber);
-        Account target = accountCommandRepository.findByAccountNumber(targetAccountNumber);
-        if (source == null || target == null) {
-            throw new IllegalArgumentException("Source or target account not found");
-        }
+        Account source = accountCommandRepository.findByAccountNumber(sourceAccountNumber)
+                .orElseThrow(() -> new IllegalArgumentException("Source account not found with account number: " + sourceAccountNumber));
+        Account target = accountCommandRepository.findByAccountNumber(targetAccountNumber)
+                .orElseThrow(() -> new IllegalArgumentException("Target account not found with account number: " + targetAccountNumber));
 
         source.withdraw(amount);
         target.deposit(amount);
 
-        var mov = accountMapper.toEntity(source.getAccountNumber(), TransactionType.TRANSFERENCIA, amount, target.getAccountNumber());
-        source.getTransactions().add(mov);
+        var transaction = accountMapper.toEntity(source.getAccountNumber(), TransactionType.TRANSFERENCIA, amount, target.getAccountNumber());
+        source.getTransactions().add(transaction);
         accountCommandRepository.save(source);
         accountCommandRepository.save(target);
 
         return new TransferResponseDTO(
                 source.getAccountNumber(),
-                LocalDateTime.now(ZoneOffset.of("-03:00")).toString(),
+                OffsetDateTime.now(ZoneOffset.of("-03:00")).toString(),
                 target.getAccountNumber(),
                 source.getBalance(),
                 amount
@@ -100,11 +94,9 @@ public class AccountCommandService {
 
     }
 
-    public AccountResponseDTO setLimit(String numero, BigDecimal limite) {
-        Account account = accountCommandRepository.findByAccountNumber(numero);
-        if (account == null) {
-            throw new IllegalArgumentException("Account not found with account number: " + numero);
-        }
+    public AccountResponseDTO setLimit(String accountNumber, BigDecimal limite) {
+        Account account = accountCommandRepository.findByAccountNumber(accountNumber)
+                .orElseThrow(() -> new IllegalArgumentException("Account not found with account number: " + accountNumber));
         account.setLimitAmount(limite);
         accountCommandRepository.save(account);
