+++ bank-account/src/main/java/com/dac/bank_account/enums/TransactionType.java
@@ -1,4 +1,4 @@
-package com.dac.bank_account.command.enums;
+package com.dac.bank_account.enums;
 
 public enum TransactionType {
     DEPOSITO,
